.mcp/brainstorm_with_llms/pyproject.toml
[project]
name = "brainstorm-with-llms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==2.10.5",
    "google-genai>=1.27.0",
    "openai>=1.97.1",
    "pydantic>=2.11.7",
]